Move ILobbyGameBroadcaster to common package
