+++ lobby/src/main/java/games/strategy/engine/lobby/server/LobbyGameController.java
@@ -9,6 +9,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
+import games.strategy.engine.lobby.common.ILobbyGameBroadcaster;
 import games.strategy.engine.message.IRemoteMessenger;
 import games.strategy.engine.message.MessageContext;
 import games.strategy.net.GUID;
